Ported zip.py to C++ (zip.cpp).
Ported the entire file/ to C++ for future project implementation.
zip.cpp is not the final implementation version.

If an error occurs regarding the libzip library, call it directly from the include/ file (macOS): <zip.h> -> "/opt/homebrew/opt/libzip/include/zip.h".

In main(), there is an example of using zip.cpp — assign a file to the source_path variable to compress it into a .zip and save it in the project's send/ directory (can be changed).