Added builder for iOS
The project now includes a build file for compiling on iOS,
without the previous conflict where launching the app would redirect to the zlib library and attempt to run it as the main executable.

A build file for all planned platforms will be added soon.

Removed an unnecessary file call in `src/file/tozip`.