Added basic context for tabs
wxPanelForTab::InitTabContents() calls the initialization of each tab (Settings, Send, Help).
If wxPanelForTab->tabName is not found, wxPanelForTab::InitUnknownTabContents() is initialized.

The context does not represent anything, it is just a test for further improvement.

If you want to set your own tab name, don't forget to change the name in wxPanelForTab::InitTabContents().